Home

consistent_sampler.jl Documentation

Ticket

Ticket — Type.

Ticket

A struct which stores a ticket number, id, and generation

copyTicket — Method.

copyTicket(t::Ticket)

Returns a copy of the ticket 't'

Base.println — Method.

println(t::Ticket)

Prints out the values of a ticket on a line

Base.string — Method.

string(t::Ticket)

Returns a string version of a Ticket

nextTicket — Method.

nextTicket(t::Ticket)

Returns a ticket with 1 added to the generation, and the ticketnumber being set to nextfraction of the current ticket number

make_ticket_priority_queue — Method.

make_ticket_priority_queue(idlist::AbstractArray,seed)

Returns a priority queue with the values being tickets made with an id from idlist and a seed of seed, and their corresponding keys being their ticket_number

Formatting

trim — Method.

trim(x::String,length::Int=9)

Trims a string 'x' to 'length' (defaulted to 9) digits after the last consecutive 9 after decimal

Duplicates — Method.

Duplicates(L::AbstractArray)

Returns an array of all the duplicates in the array L

Hashing

sha256_hex — Method.

sha256_hex(h)

Returns the hexadecimal form of the sha256(h)

sha256_uniform — Method.

sha256_uniform(h)

Returns a decimal point followed by the string form of the digits of the decimal form of sha256_uniform(h) in reverse order

first_fraction — Method.

first_fraction(id,seed,seed_hash::String="unknown")

Returns the sha256_uniform of the id and sha256_uniform of the seed, which can also be directly implemented as seed_hash to save time for repetetive use

next_fraction — Method.

next_fraction(x::String)

Returns the first sha256_uniform of a String followed by an increasing number which is greater than the original String

With/Without Replacement

withoutReplacement — Method.

withoutReplacement(t::Array{Ticket},l::Int)

Returns the first l indices of T

withReplacement — Method.

withReplacement(t::Array{Ticket},l::Int)

Returns an array of l tickets, but every time a ticket is put in this array, its value is put back into t as nextTicket of itself

Sampler

sampler — Method.

sampler(ids::AbstractArray,; seed::Any=nothing, take::Int=length(ids),with_replacement=false,drop::Int=0,output="tuple",digits::Int=9)

Returns the withoutReplacement or withReplacement function of a priority queue of tickets, made by make_ticket_priority_queue(ids,seed), with length of drop+take, with the first drop idexes removed, outputted in output form (Ticket,tuple, or ids), with the ticket_numbers trimmed to 'digits' digits